split Slides
